requirements.txt
flask
flask_sqlalchemy
python-barcode
pillow